Make update_status utility function
